util: Move hashing / permutation functions
